Allow lookup to prioritize registry or resolver
Setting a custom resolver for the container currently makes it
difficult to use in test contexts, where we may want test
doubles for some factories in the container. They can't be
registered normally, because they resolver will ignore them.

This change preserves the existing API, but allows a
"fallbackResolver" to be set instead, which will only be used
if nothing is found in the registry.